Revert "GCI-142 Remove redundant value unboxing (#2398)"
This reverts commit 129cb11bb2a998944f773bf2113fa2934d2798dd.

+++ api/src/main/java/org/openmrs/Form.java
@@ -147,8 +147,8 @@ public class Form extends BaseChangeableOpenmrsMetadata {
 				
 				for (FormField ff : fieldSet) {
 					if (ff.getFieldNumber() != null) {
-						if (ff.getFieldNumber() < fieldNum || fieldNum == 0) {
-							fieldNum = ff.getFieldNumber();
+						if (ff.getFieldNumber().intValue() < fieldNum || fieldNum == 0) {
+							fieldNum = ff.getFieldNumber().intValue();
 							next = ff;
 						}
 					} else {
